Reflex.Material.Common

Documentation

class MdClassText a where

A type class for converting data types into appropriate MDC class text.

Minimal complete definition

mdText

Methods

mdText :: a -> Text

class MdHasCustom a where

Minimal complete definition

custom

Methods

custom :: Text -> a -> a

IMPORTANT: Implementations of this function should use the accompanying addCustom function to make sure that new values are added on and don't overwrite anything that was already there.

addCustom :: Text -> Maybe Text -> Maybe Text

Helper function for adding a class item to a custom class field.

data MdRole

Constructors

MdPrimary

MdAccent

class MdHasRole a where

Minimal complete definition

mdSetRole

Methods

mdSetRole :: MdRole -> a -> a

primary :: MdHasRole a => a -> a

accent :: MdHasRole a => a -> a

data MdRaised

Constructors

MdRaised

class MdHasRaised a where

Minimal complete definition

raised

Methods

raised :: a -> a

data MdDark

Constructors

MdDark

class MdHasDark a where

Minimal complete definition

dark

Methods

dark :: a -> a

data MdActive

Constructors

MdActive

class MdHasActive a where

Minimal complete definition

active

Methods

active :: a -> a

data MdDisabled

Constructors

MdDisabled

class MdHasDisabled a where

Minimal complete definition

disabled

Methods

disabled :: a -> a

data MdSize

Constructors

MdMini

MdTiny

MdSmall

MdMedium

MdLarge

MdBig

MdHuge

MdMassive

class MdHasSize a where

Minimal complete definition

mdSetSize

Methods

mdSetSize :: MdSize -> a -> a

mini :: MdHasSize a => a -> a

tiny :: MdHasSize a => a -> a

small :: MdHasSize a => a -> a

medium :: MdHasSize a => a -> a

large :: MdHasSize a => a -> a

big :: MdHasSize a => a -> a

huge :: MdHasSize a => a -> a

massive :: MdHasSize a => a -> a

data MdDensity

Constructors

MdCompact

MdDense

class MdHasDensity a where

Minimal complete definition

mdSetDensity

Methods

mdSetDensity :: MdDensity -> a -> a

compact :: MdHasDensity a => a -> a

dense :: MdHasDensity a => a -> a

tshow :: Show a => a -> Text