Reflex.Material.Button

Documentation

mdButton :: MonadWidget t m => Dynamic t MdButton -> m () -> m (Event t ())

Buttons are for clicking

data MdButton

Data structure describing options available for buttons. The typical way of using this data structure is to use the default instance and modify it using the various MdHasXYZ type classes. For instance:

huge $ inverted $ blue def

Constructors

MdButton

Fields

_mdButton_role :: Maybe MdRole

_mdButton_raised :: Maybe MdRaised

_mdButton_activation :: Maybe (Either MdActive MdDisabled)

active and disabled should be mutually exclusive so we use an Either

_mdButton_size :: Maybe MdSize

_mdButton_density :: Maybe MdDensity

_mdButton_custom :: Maybe Text